Define modular structure for the crates
