+++ .github/workflows/benchmark.yml
@@ -1,0 +1,19 @@
+name: Run benchmarks
+
+on:
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  run_benchmarks:
+    name: Run Benchmarks
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - uses: boa-dev/criterion-compare-action@v3
+        with:
+          branchName: ${{ github.base_ref }}
+          token: ${{ secrets.GITHUB_TOKEN }}
