+++ .github/workflows/build.yml
@@ -1,0 +1,36 @@
+name: Verify build
+
+on: [push]
+
+env:
+  CARGO_TERM_COLOR: always
+
+jobs:
+  check:
+    name: Build ArkLib
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install Rust
+        uses: dtolnay/rust-toolchain@stable
+        with:
+            components: rustfmt, clippy
+
+      - name: Check
+        run: cargo check
+
+      - name: Format
+        run: |
+          cargo fmt --all -- --check
+          cargo clippy 
+
+      - name: Build Debug
+        run: cargo build --verbose
+
+      - name: Run tests
+        run: cargo test --verbose
+
+      - name: Build Release
+        run: cargo build --verbose --release
