+++ crates/fs-atomic-versions/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "fs-atomic-versions"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "fs_atomic_versions"
+crate-type = ["rlib"]
+bench = false
+
+[dependencies]
+fs-utils = { path = "../fs-utils" }
+
+log = { version = "0.4.17", features = ["release_max_level_off"] }
+anyhow = "1.0.58"
+lazy_static = "1.4.0"
+serde_json = "1.0.82"
+serde = { version = "1.0.138", features = ["derive"] }
+fastrand = "2"
+uuid = { version = "1.6.1", features = ["v4"] }
+
+[dev-dependencies]
+tempdir = "0.3.7"
+rstest = '0.18.2'
