+++ crates/fs-index/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "fs-index"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "fs_index"
+crate-type = ["rlib"]
+bench = false
+
+[dependencies]
+fs-utils = { path = "../fs-utils" }
+fs-atomic-versions = { path = "../fs-atomic-versions" }
+
+log = { version = "0.4.17", features = ["release_max_level_off"] }
+crc32fast = "1.3.2"
+walkdir = "2.3.2"
+anyhow = "1.0.58"
+lazy_static = "1.4.0"
+canonical-path = "2.0.2"
+pathdiff = "0.2.1"
+serde_json = "1.0.82"
+serde = { version = "1.0.138", features = ["derive"] }
+url = { version = "2.2.2", features = ["serde"] }
+reqwest = "0.11.11"
+scraper = "0.13.0"
+tokio = { version = "1", features = ["full"] }
+itertools = "0.10.5"
+
+[dev-dependencies]
+uuid = { version = "1.6.1", features = ["v4"] }
+tempdir = "0.3.7"
+rstest = '0.18.2'
+# benchmarking
+criterion = { version = "0.5", features = ["html_reports"] }
+pprof = { version = "0.13", features = ["criterion", "flamegraph"] }
+rand = "0.8"
+
+[build-dependencies]
+flate2 = "1.0.24"
+fs_extra = "1.2.0"
+tar = "0.4.38"
+target-lexicon = "0.12.4"
+ureq = "2.4.0"
+ring = "=0.17.5"
+
+[[bench]]
+name = "compute_bytes_benchmark"
+harness = false
+path = "benches/compute_bytes_benchmark.rs"
+
+[[bench]]
+name = "index_build_benchmark"
+harness = false
+path = "benches/index_build_benchmark.rs"
