+++ crates/fs-index/src/storage/mod.rs
@@ -1,0 +1,2 @@
+pub mod meta;
+pub mod prop;
