+++ crates/fs-utils/src/errors.rs
