--- fs-atomic-versions/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "fs-atomic-versions"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "fs-atomic-versions"
-crate-type = ["rlib"]
-bench = false
-
-[dependencies]
-log = { version = "0.4.17", features = ["release_max_level_off"] }
-anyhow = "1.0.58"
-env_logger = "0.9.0"
-lazy_static = "1.4.0"
-canonical-path = "2.0.2"
-pathdiff = "0.2.1"
-image = "0.24.2"
-libloading = "0.7.3"
-serde_json = "1.0.82"
-serde = { version = "1.0.138", features = ["derive"] }
-url = { version = "2.2.2", features = ["serde"] }
-reqwest = "0.11.11"
-scraper = "0.13.0"
-zip = "0.6.2"
-tokio = { version = "1", features = ["full"] }
-itertools = "0.10.5"
-once_cell = "1.16.0"
-thiserror = "1"
-fastrand = "2"
-uuid = { version = "1.6.1", features = ["v4"] }
-
-[dev-dependencies]
-tempdir = "0.3.7"
-rstest = '0.18.2'
-
-[build-dependencies]
-flate2 = "1.0.24"
-fs_extra = "1.2.0"
-libloading = "0.7.3"
-tar = "0.4.38"
-target-lexicon = "0.12.4"
-ureq = "2.4.0"
-ring = "=0.17.5"
